Add propagate-uid-gid option
Add an option, `propagate-uid-gid`, to match the container UID/GID to
the host UID/GID. This avoids littering shared mounts with root-owned
files. The existing `user` option isn't quite flexible enough, as it
requires you to know the name of the user you want to run as, and often
you just want to use whatever UID/GID the host agent happens to be
running as.